telegram-service/requirements.txt
kafka-python==2.0.2
httpx==0.27.0
python-dotenv==1.0.1
tenacity==8.3.0
